Remove ForeignKey requirement of the neighbor AS from RouterWeb.
- This PR Fixes https://github.com/netsec-ethz/scion-web/issues/65
- It removes the ForeignKey requirement of the neighbor AS from RouterWeb model
since local scion-web instances in the testbed will not have the neighbor ASes in their
local databases in most cases. Instead, the neighbor AS is represented with its ISD and
AS IDs.
- Adapts the existing network view (visualization) tab: If the neighboring AS is not in
the local scion-web database, then it is not shown in the network view. Only the network
view of the ASes that are known to the local scion-web instance are shown.
- Adapts the tests accordingly.
- Where applicable, renames 'ad' to 'as' in the variables used.